1st test Web3modal integration
