--- app/about/page.tsx
@@ -1,10 +1,0 @@
-import About from '../components/how_to/About'
-
-export default function Home() {
-    return (
-      <>
-        <About/>
-      </>
-  
-    )
-  }
