--- app/components/ens/GetENS.tsx
@@ -1,1 +1,0 @@
-
