--- app/components/web3modal/Web3Modal.jsx
@@ -1,57 +1,0 @@
-import { createWeb3Modal } from '@web3modal/wagmi/react'
-
-import { http, createConfig, WagmiProvider } from 'wagmi'
-import { mainnet, sepolia, goerli } from 'wagmi/chains'
-import { walletConnect, injected, coinbaseWallet } from 'wagmi/connectors'
-
-import { QueryClient, QueryClientProvider } from '@tanstack/react-query'
-
-// 0. Setup queryClient
-const queryClient = new QueryClient()
-
-// 1. Get projectId at https://cloud.walletconnect.com
-const projectId = 'efc3c7fc3d9dab0817cdead6a77b09f8'
-
-// 2. Create wagmiConfig
-const metadata = {
-  name: 'Web3Modal',
-  description: 'Web3Modal QuoteApp',
-  url: 'https://my-quote-app-react.vercel.app/',//'http://localhost:3000',//'https://web3modal.com', // origin must match your domain & subdomain
-  icons: ['https://avatars.githubusercontent.com/u/37784886']
-}
-
-const config = createConfig({
-  chains: [mainnet, sepolia, goerli],
-  transports: {
-    [mainnet.id]: http(),
-    [sepolia.id]: http(),
-    [goerli.id]: http()
-  },
-  connectors: [
-    walletConnect({ projectId, metadata, showQrModal: false }),
-    injected({ shimDisconnect: true }),
-    coinbaseWallet({
-      appName: metadata.name,
-      appLogoUrl: metadata.icons[0]
-    })
-  ]
-})
-
-// 3. Create modal
-createWeb3Modal({
-  wagmiConfig: config,
-  projectId,
-  enableAnalytics: true // Optional - defaults to your Cloud configuration
-})
-
-export function ContextProvider({ children }) {
-  return (
-    <WagmiProvider config={config}>
-      <QueryClientProvider client={queryClient}>{children}</QueryClientProvider>
-    </WagmiProvider>
-  )
-}
-
-export default function ConnectButton() {
-    return <w3m-button />
-  }
