--- next.config.js
@@ -1,9 +1,0 @@
-/** @type {import('next').NextConfig} */
-const nextConfig = {
-    reactStrictMode: true,
-    // experimental: {
-    //     serverActions: true,
-    //   },
-}
-
-module.exports = nextConfig
