+++ package.json
@@ -1,43 +1,35 @@
 {
-  "name": "web3js",
+  "name": "My Quote React App",
   "version": "0.1.0",
   "private": true,
   "scripts": {
     "dev": "next dev",
-    "info": "next info",
     "build": "next build",
     "start": "next start",
     "lint": "next lint"
   },
   "dependencies": {
-    "@metamask/detect-provider": "^2.0.0",
-    "@tanstack/react-query": "^5.25.0",
-    "@types/react-helmet": "^6.1.11",
-    "@types/react-select": "^5.0.1",
-    "@web3modal/wagmi": "^4.0.13",
-    "bootstrap": "^5.3.2",
-    "bufferutil": "^4.0.8",
-    "encoding": "^0.1.13",
-    "next": "^14.0.5-canary.41",
+    "@tanstack/react-query": "^5.28.0",
+    "@web3modal/wagmi": "^4.1.1",
+    "next": "^14.1.3",
     "react": "^18.2.0",
-    "react-bootstrap": "^2.9.1",
+    "react-bootstrap": "^2.10.1",
     "react-dom": "^18.2.0",
     "react-select": "^5.8.0",
-    "sass": "^1.69.5",
-    "utf-8-validate": "^6.0.3",
-    "viem": "^2.7.19",
-    "wagmi": "^2.5.7",
-    "web3": "^4.2.0"
+    "ts-node": "^10.9.2",
+    "viem": "^2.8.6",
+    "wagmi": "^2.5.7"
   },
   "devDependencies": {
+    "@tanstack/eslint-plugin-query": "^5.27.7",
     "@types/node": "^20",
     "@types/react": "^18",
     "@types/react-dom": "^18",
-    "autoprefixer": "^10",
+    "autoprefixer": "^10.0.1",
     "eslint": "^8",
-    "eslint-config-next": "^14.0.4",
+    "eslint-config-next": "14.1.3",
     "postcss": "^8",
-    "tailwindcss": "^3",
+    "tailwindcss": "^3.3.0",
     "typescript": "^5"
   }
 }
