+++ src/app/components/button/ConnectButton.jsx
@@ -1,0 +1,3 @@
+export default function ConnectButton() {
+    return <w3m-button />
+  }
