+++ src/app/components/main/App.tsx
@@ -1,0 +1,12 @@
+import GetQuoteButton from "@/components/button/GetQuoteButton"
+import ConnectButton from "@/components/button/ConnectButton"
+
+export default function App()
+{
+    return (
+        <>
+            <ConnectButton/>
+            <GetQuoteButton/>
+        </>
+    )
+}
