+++ src/app/layout.tsx
@@ -1,0 +1,29 @@
+import '@/css/globals.css'
+import type { Metadata } from 'next'
+import { headers } from 'next/headers'
+
+import { cookieToInitialState } from 'wagmi'
+
+import { config } from '@/config'
+import { ContextProvider } from '@/context'
+
+export const metadata: Metadata = {
+  title: 'Create Next App',
+  description: 'Generated by create next app'
+}
+
+export default function RootLayout({
+  children
+}: Readonly<{
+  children: React.ReactNode
+}>) {
+  const initialState = cookieToInitialState(config, headers().get('cookie'))
+  return (
+    <html lang="en">
+      <body>
+        <ContextProvider initialState={initialState}>{children}</ContextProvider>
+        {/* {children} */}
+      </body>
+    </html>
+  )
+}
