+++ src/app/page.tsx
@@ -1,0 +1,10 @@
+import Image from "next/image";
+import App from '@/components/main/App'
+
+export default function Home() {
+  return (
+    <main >
+      <App/>
+    </main>
+  );
+}
